Move PageLoading from form to page, to prevent duplicate headers
